requirements.txt
numpy
scipy
torch>=2.1.2
torchvision
git+https://github.com/NVlabs/nvdiffrast
ninja
# common
tqdm
calibur
simple-parsing
# mesh
trimesh>=4.1.3
pycollada
pyglet<2
# visualization
pillow
matplotlib
opencv-python
# dev
ipykernel